requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain-google-genai>=1.0.0
tqdm>=4.65.0
setuptools>=65.0.0

# GCP Vertex AI (for higher rate limits with credits)
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Audio processing
gTTS>=2.3.0
pydub>=0.25.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0,<2.0.0

# Optional: Enhanced TTS providers
requests>=2.32.0